# Optimize `URI.decode`